functions/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
# azure-monitor-opentelemetry

azure-functions
azure-cosmos>=4.5.0
requests>=2.31.0
azure-storage-blob>=12.23.0
requests-toolbelt>=1.0.0
openai
azurefunctions-extensions-bindings-blob>=1.1.1
